pyproject.toml
[project]
name = "cloud-guardrails-bot"
version = "0.4.0"
description = "Policy-driven FinOps & Security remediation via Slack approvals (OTP + dual approval for prod)"
requires-python = ">=3.11"
authors = [{name="Wellington Teixeira"}]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "redis>=5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "bandit>=1.7",
  "pip-audit>=2.7",
  "types-python-dateutil",
]

[tool.ruff]
line-length = 120
target-version = "py311"
fix = false

[tool.ruff.lint]
select = ["E","F","I","B","UP","SIM","S"]
ignore = ["S101","E501"]  # allow asserts in tests; line length enforced by formatter # allow asserts in tests

[tool.ruff.lint.per-file-ignores]
"src/guardrails/tests/*" = ["S105","S106","S107"]

[tool.pytest.ini_options]
testpaths = ["src/guardrails/tests"]


[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
